Réorganisation modulaire du code
