+++ Start.py
@@ -1,0 +1,4 @@
+from LOGICIEL import main
+
+if __name__ == "__main__":
+    main()
